backend/pyproject.toml
[tool.poetry]
name = "bellweaver"
version = "0.1.0"
description = "School calendar event aggregation and filtering tool"
authors = ["Jethro <jethro@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
requests = "^2.32.5"
cryptography = "^46.0.3"
anthropic = "^0.74.1"
sqlalchemy = "^2.0.44"
flask = "^3.1.2"
python-dotenv = "^1.2.1"
beautifulsoup4 = "^4.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.0.0"
pyppeteer = "^1.0.2"
pyppeteer-stealth = "^2.7.4"

[tool.poetry.scripts]
bellweaver = "src.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
